+++ src/views/auth/EnterOTP.vue
@@ -2,7 +2,7 @@
     <OTP></OTP>
 </template>
 <script>
-import OTP from "@/components/OTP.vue";
+import OTP from "@/components/auth/OTP.vue";
 
 export default {
   name: "EnterOTP",
